hw1/requirements.txt
swig
gymnasium[mujoco,box2d]
torch
numpy
tensorboard
tensorboardX
matplotlib
ipython
moviepy
pyvirtualdisplay
opencv-python
ipdb
